+++ release.config.js
@@ -1,6 +1,17 @@
 module.exports = {
     branches: [
         {
+            name: "next",
+            prerelease: true,
+            channel: "next",
+            level: "major",
+            devDependencies: {
+                "@zowe/cli": "next",
+                "@zowe/imperative": "next",
+                "@zowe/cli-test-utils": "next"
+            }
+        },
+        {
             name: "master",
             level: "minor",
             devDependencies: ["@zowe/cli", "@zowe/imperative", "@zowe/cli-test-utils"]
@@ -10,17 +21,12 @@ module.exports = {
             level: "patch",
             devDependencies: ["@zowe/cli", "@zowe/imperative"]
         }
-        // {
-        //     name: "next",
-        //     prerelease: true,
-        //     devDependencies: ["@zowe/cli", "@zowe/imperative", "@zowe/cli-test-utils"]
-        // }
     ],
     plugins: [
         "@octorelease/changelog",
         ["@octorelease/npm", {
             aliasTags: {
-                latest: ["zowe-v2-lts", "next"]
+                latest: ["zowe-v2-lts"]
             },
             pruneShrinkwrap: true
         }],